Implement configuration file module
Ideally I would like to use TOML or YAML instead to have better diffs
here. However, I haven't been able to find libraries that are as
understandable or consistent as encoding/json.

When passing in a type that implements `Marshal[TYPE]` or
`Unmarshal[TYPE]`, I would expect those methods to be called. However,
it seems like the libaries I've tried to use for both TOML and YAML
only call these methods as a secondary measure, and so my `map` type is
incorrectly serialized.

These are the libraries I tried:
- gopkg.in/yaml.v2
- github.com/pelletier/go-toml
- github.com/BurntSushi/toml
- github.com/naoina/toml

:(