Knoch files can live outside current dir
Knoch files will now work similarly to .vimrc or .git

Knoch will look for .knoch files going upwards until it hits either the
home directory or root. If it doesn't find any files, it will assume
that the current directory is the working directory.

All directories stored in the knoch file will be relative to it -
regardless of how the user inputs a directory for knoch to track.