Fix bug related to `knoch remove` with completion
Our completion gives names of projects relative to the .knoch directory.
The remove command, up to now, was only trying to remove a project
relative to the current working directory, not the .knoch directory.
This means that any time you used zsh_completion, remove would fail
unless you happened to be in the .knoch directory.

We now fall back to checking a directory relative to the .knoch
directory when it doesn't exist locally.